\name{hbim-package}
\alias{hbim-package}
\alias{hbim}
\docType{package}
\title{Hill/Bliss Independence Model for Multicomponent Vaccines}
\description{Calculate expected relative risk and proportion protected assuming normally distributed log10 transformed antibody dose for several component vaccine. Uses Hill models for each component which are combined under Bliss independence.  
}
\details{
\tabular{ll}{
Package: \tab hbim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-13\cr
License: \tab GPL \cr
}
The hbim package allows users to reproduce plots and calculations for Saul and Fay (2007). See \code{vignette("hbimdetails")}.
}
\author{
M.P. Fay, 
Maintainer: Michael Fay <mfay@niaid.nih.gov>
}
\references{
Saul A, Fay MP (2007).  Human Immunity and the Design of Multi-Component, Single Target Vaccines. PLoS ONE 2(9): e850. doi:10.1371/jounal.pone.0000850
}
\keyword{ package }
%\seealso{~~ Optional links to other man pages, e.g. ~~~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~}
%\examples{
%~~ simple examples of the most important functions ~~
%}
