\name{refs}
\alias{refs}
\docType{data}
\title{Reference list}
\description{Each reference is one long character string. See \code{data(irdata)} for data from each reference. 
}
\usage{data(refs)}
\format{
  The format is:
 Factor w/ 50 levels (the 50 references)
}
%\details{Each reference is one long character string.}
%\source{  ~~ reference to a publication or URL from which the data were obtained ~~}
%\references{ ~~ possibly secondary sources and usages ~~}
\examples{
data(refs)
refs[1]
}
\keyword{datasets}
